# Conflicts:
#	spring-batch/pom.xml

+++ persistence-modules/spring-data-jpa-enterprise-2/src/test/java/com/baeldung/namingstrategy/QuotedUpperCaseNamingStrategyPostgresLiveTest.java
@@ -3,14 +3,9 @@ package com.baeldung.namingstrategy;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
 
 import org.hibernate.exception.SQLGrammarException;
 import org.junit.jupiter.api.BeforeEach;
@@ -21,6 +16,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.TestDatabaseAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.TestPropertySource;
+
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import jakarta.persistence.Query;
 
 @DataJpaTest(excludeAutoConfiguration = TestDatabaseAutoConfiguration.class)
 @TestPropertySource("quoted-upper-case-naming-strategy-on-postgres.properties")
@@ -73,8 +72,7 @@ class QuotedUpperCaseNamingStrategyPostgresLiveTest {
     public Person fromDatabase(Object databaseRow) {
         Object[] typedDatabaseRow = (Object[]) databaseRow;
 
-        return new Person(
-          ((BigInteger) typedDatabaseRow[0]).longValue(),
+        return new Person((Long) typedDatabaseRow[0],
           (String) typedDatabaseRow[1],
           (String) typedDatabaseRow[2]
         );
